docker-images/vulnerable-api/requirements.txt
flask==1.1.1
requests==2.20.0
pyyaml==5.1

